Multi-machine refactor phase 3 (part 5): collapse GLUE state into one struct
Replace the separate se30_state_t and iicx_state_t with a single
mac030_glue_state_t (proposal §4.2.3 "the three _internal.h clone headers
collapse into one substrate state struct"). It is a superset: the SE/30 uses
vram/vrom/video_card and leaves the IIcx soft-power members unused; IIcx/IIx
use last_via2_port_b/soft_power_armed and leave the video pointers NULL.

se30_state_t and iicx_state_t become typedef aliases of the unified struct;
the se30_state()/iicx_state() accessors and every field reference are
unchanged (field names already matched). No behaviour change — pure type
unification; the struct is never serialised raw (checkpoints save individual
fields), so layout is irrelevant.

Verified: se30-boot, iicx-mactest, iix-boot, iici-boot, iisi-boot pass.
